changes again

+++ convert_ratings_to_csv.py
@@ -13,7 +13,7 @@ from system import *
 
 def main(params=[]):
     parser = argparse.ArgumentParser(description='avrate++ convert ratings to csv for better handling', epilog="stg7 2017", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('dbfilename', type=str, help='filename of database with ratings')
+    parser.add_argument('--dbfilename', type=str, default="ratings.db", help='filename of database with ratings')
     parser.add_argument('--cvsfilename', type=str, default="ratings.csv", help='filename of cvs file for exporting')
 
     argsdict = vars(parser.parse_args())
@@ -21,16 +21,19 @@ def main(params=[]):
 
     connection = sqlite3.connect(argsdict["dbfilename"])
 
+    if not os.path.isfile(argsdict["dbfilename"]):
+        lError("your database is not a valid file")
+        return
+
     schema = {}
     for row in connection.execute("""pragma table_info('ratings') """):
         schema[row[1]] = ""
     schema = sorted(schema.keys())
-    print(schema)
+
     with open(argsdict["cvsfilename"], "w") as csv:
         csv.write(";".join(schema) + "\n")
         for row in connection.execute("""select {} from ratings; """.format(",".join(schema))):
             csv_values = [str(row[i]) for i in range(len(row))]
-            #csv_values[schema.index("rating_filled")] = str(row[schema.index("rating_filled")]).split("=")[1].replace("'", "")
             csv.write(";".join(csv_values) + "\n")
 
     connection.commit()
